+++ packages/core/src/crypto/jwt.test.ts
@@ -13,9 +13,9 @@ const streamId = 'd7ae985a-7a27-4c18-a00c-22a5b5ea7e10';
 
 const commonOptions = {
   privateKey:
-    '-----BEGIN PRIVATE KEY-----\nMIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQgIiR3kYwhLTMY7Jfe\nZZ20lK5WF8QcUhiBfr/VXukkijGhRANCAASijOcHY3nUo9ZiA1Z9JK+UMKAtHM+3\nvD987L87T7E59y49z+7D6nruo9TAm+vhr7AuCBy8OoalKKoiKcChO0Y4\n-----END PRIVATE KEY-----\n',
+    'LS0tLS1CRUdJTiBQUklWQVRFIEtFWS0tLS0tCk1JR0hBZ0VBTUJNR0J5cUdTTTQ5QWdFR0NDcUdTTTQ5QXdFSEJHMHdhd0lCQVFRZzlkeWpDWVI4RWl0UzBqUUoKQkNxbUJqMXdxVFhWeVZ6L3NCQmQ3WkxVcUIyaFJBTkNBQVRKQVpKVjN3Z3F5RkMwNk1qajNIc1NVbks0Z0RyZQpFMjFNNU8xWFRjNFM0dHdJbkwxV29zRi9JLzAzblhQQU5lVEV3dmk0dWpnL0FHNzg5VEk2UFJPMAotLS0tLUVORCBQUklWQVRFIEtFWS0tLS0tCg==',
   publicKey:
-    'LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFb296bkIyTjUxS1BXWWdOV2ZTU3ZsRENnTFJ6UAp0N3cvZk95L08wK3hPZmN1UGMvdXcrcDY3cVBVd0p2cjRhK3dMZ2djdkRxR3BTaXFJaW5Bb1R0R09BPT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==',
+    'LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFeVFHU1ZkOElLc2hRdE9qSTQ5eDdFbEp5dUlBNgozaE50VE9UdFYwM09FdUxjQ0p5OVZxTEJmeVA5TjUxendEWGt4TUw0dUxvNFB3QnUvUFV5T2owVHRBPT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==',
   issuer: 'https://livepeerjs.org',
 } as const;
 
@@ -45,9 +45,7 @@ describe('signAccessJwt', () => {
       }
     `);
     expect(decoded.payload.action).toEqual('pull');
-    expect(decoded.payload.pub).toEqual(
-      'LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFb296bkIyTjUxS1BXWWdOV2ZTU3ZsRENnTFJ6UAp0N3cvZk95L08wK3hPZmN1UGMvdXcrcDY3cVBVd0p2cjRhK3dMZ2djdkRxR3BTaXFJaW5Bb1R0R09BPT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==',
-    );
+    expect(decoded.payload.pub).toEqual(commonOptions.publicKey);
     expect(decoded.payload.iss).toEqual('https://livepeerjs.org');
     expect(decoded.payload.sub).toEqual('abcd1234');
     expect(decoded.payload.video).toEqual('none');
@@ -84,9 +82,7 @@ describe('signAccessJwt', () => {
       }
     `);
     expect(decoded.payload.action).toEqual('pull');
-    expect(decoded.payload.pub).toEqual(
-      'LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFb296bkIyTjUxS1BXWWdOV2ZTU3ZsRENnTFJ6UAp0N3cvZk95L08wK3hPZmN1UGMvdXcrcDY3cVBVd0p2cjRhK3dMZ2djdkRxR3BTaXFJaW5Bb1R0R09BPT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==',
-    );
+    expect(decoded.payload.pub).toEqual(commonOptions.publicKey);
     expect(decoded.payload.iss).toEqual('https://livepeerjs.org');
     expect(decoded.payload.sub).toEqual('d7aer9qx8act4lfd');
     expect(decoded.payload.video).toEqual('none');
@@ -109,9 +105,7 @@ describe('signAccessJwt', () => {
       }
     `);
     expect(decoded.payload.action).toEqual('pull');
-    expect(decoded.payload.pub).toEqual(
-      'LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUZrd0V3WUhLb1pJemowQ0FRWUlLb1pJemowREFRY0RRZ0FFb296bkIyTjUxS1BXWWdOV2ZTU3ZsRENnTFJ6UAp0N3cvZk95L08wK3hPZmN1UGMvdXcrcDY3cVBVd0p2cjRhK3dMZ2djdkRxR3BTaXFJaW5Bb1R0R09BPT0KLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg==',
-    );
+    expect(decoded.payload.pub).toEqual(commonOptions.publicKey);
     expect(decoded.payload.iss).toEqual('https://livepeerjs.org');
     expect(decoded.payload.sub).toEqual('d7aer9qx8act4lfd');
     expect(decoded.payload.video).toEqual('none');