+++ packages/core/src/utils/index.ts
@@ -1,5 +1,5 @@
-export { b64UrlDecode, b64UrlEncode } from './b64';
 export { deepMerge } from './deepMerge';
 export { pick } from './pick';
+export { b64Decode, b64Encode, b64UrlDecode, b64UrlEncode } from './string';
 export * from './types';
 export { warn } from './warn';