+++ packages/react/src/components/media/controls/ControlsContainer.tsx
@@ -1,6 +1,5 @@
 import { MediaControllerState } from 'livepeer/media/controls';
 import { styling } from 'livepeer/styling';
-import { isNumber } from 'livepeer/utils';
 import * as React from 'react';
 
 import { useMediaController } from '../context';
@@ -20,7 +19,7 @@ const mediaControllerSelector = ({
 });
 
 export type ControlsContainerProps = {
-  uploadProgress?: number;
+  topLoadingText?: string | null;
   showLoadingSpinner?: boolean;
   hidePosterOnPlayed?: boolean;
   poster?: React.ReactNode;
@@ -43,7 +42,7 @@ export const ControlsContainer = React.forwardRef<
     poster,
     showLoadingSpinner = true,
     hidePosterOnPlayed = true,
-    uploadProgress,
+    topLoadingText,
   } = props;
 
   const { hidden, togglePlay, canPlay, hasPlayed, buffered } =
@@ -59,14 +58,6 @@ export const ControlsContainer = React.forwardRef<
       togglePlay();
     }
   }, [togglePlay, isLoaded]);
-
-  const loadingText = React.useMemo(
-    () =>
-      isNumber(uploadProgress)
-        ? `Processing: ${(Number(uploadProgress) * 100).toFixed(0)}%`
-        : null,
-    [uploadProgress],
-  );
 
   return (
     <>
@@ -92,9 +83,11 @@ export const ControlsContainer = React.forwardRef<
           className={styling.controlsContainer.background()}
           onMouseUp={onClickBackground}
         >
-          <div className={styling.controlsContainer.loadingText()}>
-            {loadingText}
-          </div>
+          {topLoadingText && (
+            <div className={styling.controlsContainer.loadingText()}>
+              {topLoadingText}
+            </div>
+          )}
 
           <div className={styling.controlsContainer.loading()} />
         </div>