+++ packages/react/src/components/media/usePlaybackInfoOrImport.tsx
@@ -7,7 +7,7 @@ import { useAsset, useCreateAsset, usePlaybackInfo } from '../../hooks';
 import { PlayerProps } from './Player';
 
 export type UsePlaybackInfoOrImportProps = {
-  src: PlayerProps['src'];
+  decentralizedSrcOrPlaybackId: ReturnType<typeof parseCid>;
   playbackId: PlayerProps['playbackId'];
   refetchPlaybackInfoInterval: number;
   autoUrlUpload: boolean;
@@ -16,13 +16,13 @@ export type UsePlaybackInfoOrImportProps = {
 
 /**
  * Retrieves the playback info for a playback ID or source URL.
- * Automatically imports a source URL from IPFS if it is a valid CID.
+ * Conditionally, automatically imports a source URL from IPFS or Arweave.
  *
  * @param src Source URL for the media.
  * @param playbackId Playback ID of the media.
  */
 export const usePlaybackInfoOrImport = ({
-  src,
+  decentralizedSrcOrPlaybackId,
   playbackId,
   refetchPlaybackInfoInterval,
   autoUrlUpload,
@@ -31,7 +31,7 @@ export const usePlaybackInfoOrImport = ({
   const { mutate: importAsset, data: importedAsset } = useCreateAsset();
 
   const { data: asset } = useAsset({
-    assetId: importedAsset?.id,
+    assetId: importedAsset?.[0]?.id,
     refetchInterval: (asset) =>
       asset?.status?.phase !== 'ready' ? refetchPlaybackInfoInterval : false,
   });
@@ -42,20 +42,9 @@ export const usePlaybackInfoOrImport = ({
     }
   }, [asset?.status, onAssetStatusChange]);
 
-  // check if the src or playbackId are IPFS sources (does not handle src arrays)
-  const ipfsSrcOrPlaybackId = React.useMemo(
-    () =>
-      playbackId
-        ? parseCid(playbackId)
-        : !Array.isArray(src)
-        ? parseCid(src)
-        : null,
-    [playbackId, src],
-  );
-
   const { data: playbackInfo, error: playbackInfoError } = usePlaybackInfo({
-    // attempt to fetch if the source is IPFS, or a playback ID is provided
-    playbackId: ipfsSrcOrPlaybackId?.cid ?? playbackId ?? undefined,
+    // attempt to fetch if the source is from decentralized storage, or a playback ID is provided
+    playbackId: decentralizedSrcOrPlaybackId?.id ?? playbackId ?? undefined,
     refetchInterval: (info) => (info ? false : refetchPlaybackInfoInterval),
   });
 
@@ -65,19 +54,23 @@ export const usePlaybackInfoOrImport = ({
     if (
       autoUrlUpload &&
       !importedAsset &&
-      ipfsSrcOrPlaybackId?.url &&
-      ipfsSrcOrPlaybackId?.cid &&
+      decentralizedSrcOrPlaybackId?.url &&
+      decentralizedSrcOrPlaybackId?.id &&
       (playbackInfoError as HttpError)?.code === 404
     ) {
       importAsset({
-        name: ipfsSrcOrPlaybackId.cid,
-        url: ipfsSrcOrPlaybackId.url,
+        sources: [
+          {
+            name: decentralizedSrcOrPlaybackId.id,
+            url: decentralizedSrcOrPlaybackId.url,
+          },
+        ],
       });
     }
   }, [
     autoUrlUpload,
     importedAsset,
-    ipfsSrcOrPlaybackId,
+    decentralizedSrcOrPlaybackId,
     playbackInfoError,
     importAsset,
   ]);